Drop CLI
There are a lot of concerns here we have to deal with that we're taking
on and I think it's unnecessary complexity. Instead, we'll go with
sticking to no CLI for now, works well with your existing test harness
(nextest, test). If we get custom test harnesses one day that'll be
cool, but I'm not sure where we sit on that as of today.